Revert "Tinder tutorial"
