+++ frontend/BarkMingle/.gitignore
@@ -33,7 +33,3 @@ yarn-error.*
 
 # typescript
 *.tsbuildinfo
-
-
-GoogleService-Info.plist
-google-services.json
