+++ frontend/BarkMingle/App.js
@@ -1,23 +1,22 @@
-import * as React from 'react';
-import StackNavigator from './StackNavigator';
-import { InteractionManager } from "react-native";
-import { NavigationContainer } from '@react-navigation/native';
-import { AuthProvider } from './hooks/useAuth';
-
+import { StatusBar } from 'expo-status-bar';
+import { StyleSheet, Text, View } from 'react-native';
 
 export default function App() {
-
-
-
   return (
-    <NavigationContainer>
-
-      <AuthProvider>
-        <StackNavigator />
-      </AuthProvider>
-
-    </NavigationContainer>
+    <View style={styles.container}>
+      <Text>The triumphant beginnings of Bark/Mingle</Text>
+      <StatusBar style="auto" />
+    </View>
   );
 }
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
