+++ frontend/BarkMingle/android/app/build.gradle
@@ -84,7 +84,6 @@ android {
     namespace 'com.anonymous.BarkMingle'
     defaultConfig {
         applicationId 'com.anonymous.BarkMingle'
-        
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
