+++ frontend/BarkMingle/babel.config.js
@@ -2,6 +2,5 @@ module.exports = function(api) {
   api.cache(true);
   return {
     presets: ['babel-preset-expo'],
-    plugins: ["nativewind/babel"],
   };
 };
