--- frontend/BarkMingle/input.css
@@ -1,1 +1,0 @@
-@tailwind utilities;
