+++ frontend/BarkMingle/package.json
@@ -6,48 +6,26 @@
     "start": "expo start",
     "android": "expo run:android",
     "ios": "expo run:ios",
-	"web": "expo start --web",
-"build:tailwind": "tailwindcss --input input.css --output tailwind.css --no-autoprefixer && tailwind-rn",
-	"dev:tailwind": "concurrently \"tailwindcss --input input.css --output tailwind.css --no-autoprefixer --watch\" \"tailwind-rn --watch\""
+    "web": "expo start --web"
   },
- "dependencies": {
-	"@expo/webpack-config": "^19.0.0",
-	"@fortawesome/fontawesome-svg-core": "^6.4.2",
-	"@fortawesome/free-solid-svg-icons": "^6.4.2",
-	"@fortawesome/react-native-fontawesome": "^0.3.0",
-	"@react-native-async-storage/async-storage": "1.18.2",
-	"@react-native-community/hooks": "^3.0.0",
-	"@react-navigation/native": "^6.1.8",
-	"@react-navigation/native-stack": "^6.9.14",
-	"@react-native-firebase/app": "^18.5.0",
-	"eas": "^0.1.0",
-	"expo": "~49.0.13",
-  "expo-auth-session": "~5.0.2",
-	"expo-crypto": "~12.4.1",
-	"expo-google-app-auth": "~8.3.0",
-	"expo-dev-client": "~2.4.11",
-  "expo-crypto": "~12.4.1",
-	"expo-google-app-auth": "~8.3.0",
-	"expo-image-picker": "~14.3.2",
-	"expo-splash-screen": "~0.20.5",
-	"expo-status-bar": "~1.6.0",
-  "expo-web-browser": "~12.3.2",
-	"firebase": "^10.5.0",
-  "nativewind": "^2.0.11",
-	"react": "18.2.0",
-	"react-dom": "18.2.0",
-	"react-native": "0.72.5",
-  "react-native-deck-swiper": "^2.0.16",
-	"react-native-safe-area-context": "4.6.3",
-	"react-native-screens": "~3.22.0",
-	"react-native-svg": "13.9.0",
-	"react-native-web": "~0.19.6",
-  "util": "^0.12.5"
+  "dependencies": {
+    "@expo/webpack-config": "^19.0.0",
+    "@react-native-firebase/app": "^18.5.0",
+    "eas": "^0.1.0",
+    "expo": "~49.0.13",
+    "expo-dev-client": "~2.4.11",
+    "expo-image-picker": "~14.3.2",
+    "expo-splash-screen": "~0.20.5",
+    "expo-status-bar": "~1.6.0",
+    "firebase": "^10.5.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-native": "0.72.5",
+    "react-native-svg": "13.9.0",
+    "react-native-web": "~0.19.6"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.0",
-    "concurrently": "^8.2.1",
-    "postcss": "^8.4.31"
+    "@babel/core": "^7.20.0"
   },
   "private": true
 }
