--- frontend/BarkMingle/styles/userProfileStyles.js
@@ -1,29 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-
-  navBar: {
-    width: "100%",
-    height: 100,
-    paddingBottom: 10,
-    backgroundColor: "#ede7d5",
-    flexDirection: "row",
-    alignContent: "center",
-    alignItems: "flex-end",
-    justifyContent: "space-evenly",
-  },
-  flex: {
-    flex: 1,
-  },
-  background: {
-    width: '100%', 
-    height: '100%',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-
-
-});
-
-
-export default styles;
