--- frontend/BarkMingle/tailwind.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  content: ["./App.{js,jsx,ts,tsx}", "./<custom directory>/**/*.{js,jsx,ts,tsx}"],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-}
