--- frontend/BarkMingle/tailwind.css
@@ -1,1 +1,0 @@
-
