--- frontend/BarkMingle/tailwind.json
@@ -1,1 +1,0 @@
-{}
